+++ CHANGELOG.md
@@ -21,3 +21,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#15] Check component health
 - [#17] add health checks before and after applying the blueprint
 - [#20] Add exception for `nginx` dependency validation. Map this dependency to `nginx-ingress` and `nginx-static`.
+- [#22] Add `dryRun` option. If `dryRun` is active the blueprint procedure stops before applying resources to the cluster and remains in the actual state. One can set the option to false and continue at this state.
+